requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
python-dotenv==1.0.0
pandas==2.1.3
numpy>=1.26.0
plotly==5.17.0
streamlit-ace==0.1.1
streamlit-option-menu==0.3.6
python-docx==0.8.11 